ws-publisher/requirements.txt
fastapi
uvicorn[standard]
google-cloud-pubsub
python-dotenv
websockets